Macros now have access to the global context. Resolves #231
